train_data_generator/basic/requirements.txt
# Core dependencies
lxml>=4.9.0
pyyaml>=6.0